--- chrome/extension/background/badge.js
@@ -1,13 +1,0 @@
-// chrome.storage.local.get('todos', (obj) => {
-//   let todos = obj.todos;
-//   if (todos) {
-//     todos = JSON.parse(todos);
-//     const len = todos.filter(todo => !todo.marked).length;
-//     if (len > 0) {
-//       chrome.browserAction.setBadgeText({ text: len.toString() });
-//     }
-//   } else {
-//     // Initial
-//     chrome.browserAction.setBadgeText({ text: '1' });
-//   }
-// });
